Split the webhook tests into files about one thing each
The old webhook test file was 799 lines. It becomes four suites — the payload's
fields, sending, registration, and renewals — sharing their fixtures through one
helpers module. Two internals that only the tests reached from outside are
private again.
